requirements.txt
# Airflow dependancies
apache-airflow==2.8.1
apache-airflow-providers-amazon
apache-airflow-providers-apache-kafka
apache-airflow-providers-apache-spark
apache-airflow-providers-celery
apache-airflow-providers-cncf-kubernetes
apache-airflow-providers-dbt-cloud
apache-airflow-providers-discord
apache-airflow-providers-docker
apache-airflow-providers-github
apache-airflow-providers-http
apache-airflow-providers-mongo
apache-airflow-providers-postgres
apache-airflow-providers-snowflake
apache-airflow-providers-sqlite
apache-airflow-providers-ssh
apache-airflow-providers-tableau
boto3
datetime
pandas
pendulum
psycopg2-binary
pymongo
requests
responses
snowflake-connector-python

faker